refactor: remove the `context` file
Integrated with the `slot_not_assigned.rs`.
