+++ kernel/src/device/pci/xhci/port/mod.rs
@@ -17,7 +17,6 @@ use slot_not_assigned::SlotNotAssigned;
 use spinning_top::Spinlock;
 
 mod class_driver;
-mod context;
 mod endpoint;
 mod resetter;
 mod slot_not_assigned;
